Choose correct token store for locations
